more methods have float args in IGraphics

+++ BaseLibS/Symbol/SymbolType.cs
@@ -35,6 +35,6 @@ namespace BaseLibS.Symbol{
 		public int Index { get; }
 		public abstract string Name { get; }
 		public abstract void GetPath(int size, out int[] pathX, out int[] pathY);
-		public abstract void Draw(int size, int x, int y, IGraphics g, Pen2 pen, Brush2 brush);
+		public abstract void Draw(int size, float x, float y, IGraphics g, Pen2 pen, Brush2 brush);
 	}
 }
